feat[do-run-in-docker]: v0.0.1

+++ lib-app.inc.bash
@@ -22,8 +22,20 @@ source "${BUILD_FUNCS_DIR%/}/lib-common.inc.bash"
 # │
 # │ function app_help()
 # │ {
-# │     echo -e "Usage: $CMD_AS_DISPLAY [--person=<name>]"
+# │     echo -e "Usage: ${COLOUR[VIVID_BLUE_USED]:-}$CMD_AS_DISPLAY [--person=<name>]${COLOUR[OFF_USED]:-}"
 # │     echo -e "       --person=<name>    Specify the name to greet (default: 'fred')"
+# │ }
+# │
+# │ function app_load_param_defaults()
+# │ {
+# │     option_person='fred'
+# │ }
+# │
+# │ function app_load_param_option_name_value()
+# │ {
+# │     [[ "$1" == '--person' ]] && option_person="$2" && return 0
+# │
+# │     return 1
 # │ }
 # │ function app_run()
 # │ {
@@ -31,32 +43,6 @@ source "${BUILD_FUNCS_DIR%/}/lib-common.inc.bash"
 # │     giveWarning "This is a warning message - these are sent to stderr"
 # │     echo "This is some more output"
 # │     echo -e "Hidden coding help is available with: ${COLOUR[VIVID_BLUE_STDOUT]:-}$CMD_AS_DISPLAY --code-help${COLOUR[OFF_STDOUT]:-}"
-# │ }
-# │ function app_load_param()
-# │ {
-# │     local load_type="$1"
-# │     local value="${2:-}"
-# │     case "$load_type" in
-# │         'all-defaults')
-# │             # set any default values for parameters here
-# │             option_person='fred'
-# │             ;;
-# │         'option-arg')
-# │             # process any options here (e.g. --my-option=value)
-# │             if [[ "$value" == '--person='* ]]; then
-# │                 option_person="${value#--person=}"
-# │             else
-# │                 return 1
-# │             fi
-# │             ;;
-# │         'direct-arg')
-# │             # process any direct arguments here (e.g. myscript.sh arg1 arg2)
-# │             return 1
-# │             ;;
-# │         *)
-# │             return 1
-# │             ;;
-# │     esac
 # │ }
 # │ # shellcheck disable=SC1091
 # │ source "$(git-shared-checkout git@github.com:smc-collaborate/ukko_bashlibs --ref="${UKKO_BASHLIBS_REF:-}")/lib-common.inc.bash"
@@ -66,7 +52,7 @@ source "${BUILD_FUNCS_DIR%/}/lib-common.inc.bash"
 #
 #    $EXE_NAME
 #    $THIS_EXE
-#    do_exit_with_help  "<optional error message>"
+#    do_errorExitWithSuggestion  "<optional error message>"
 #    do_with_check <command...>  - Run a command and exit with an error message if it fails
 #    giveWarning "<warning message>" - Give a warning message that will be shown at the end of processing (if any)
 #
@@ -76,6 +62,7 @@ function padRight()
     local width="$1"
     local text="$2"
 
+    [[ "$width" -lt 0 ]] && width=0
     printf "%-${width}s" "$text"
 }
 
@@ -90,7 +77,6 @@ function giveCodeHelp()
     echo "  * app_help()                                    | Optional but recommended"
     echo "  * app_load_param_defaults()                     | Optional but recommended"
     echo "  * app_load_param_option_name_value(name,value)  | Optional but recommended"
-    echo "  * app_load_param_option_name_only(name)         | Optional"
     echo "  * app_load_param_direct_value(value)            | Optional"
     echo "  * app_run()                                     | Required"
     echo ""
@@ -104,44 +90,66 @@ function giveCodeHelp()
     echo "    \$THIS_EXE_AS_DISPLAY │ The display path of the executable              │ ${THIS_EXE_AS_DISPLAY:-❓  Not defined}"
     echo "    \$CMD_AS_DISPLAY      │ The executable in a form that can be run        │ ${CMD_AS_DISPLAY:-❓  Not defined}"
     echo ""
-    echo "    do_exit_with_help  '<optional error message>'"
+    echo "    do_exitWithHelp  '<optional error message>'"
+    echo "    do_errorExitWithSuggestion  '<error message>'"
     echo "    do_with_check <command...>  - Run a command and exit with an error message if it fails"
     echo "    giveWarning '<warning message>' - Give a warning message that will be shown at the end of processing (if any)"
     echo "   "
 
-    coloursShow
+    colours_show
 }
 
-function do_exit_with_help()
+# shellcheck disable=SC2317
+function do_errorExitWithSuggestion()
+{
+    local fatal_error_message="${1:-}"
+    {
+        echo -e "${COLOUR[VIVID_RED_STDERR]:-}❌  $fatal_error_message${COLOUR[OFF_STDERR]:-}"
+        echo -e "Suggest: ${COLOUR[VIVID_BLUE_STDOUT]:-}$CMD_AS_DISPLAY --help${COLOUR[OFF_STDOUT]:-}"
+    } >&2
+    exit 1
+}
+
+function do_errorExitWithSuggestion()
+{
+    local fatal_error_message="${1:-}"
+    {
+        echo -e "${COLOUR[VIVID_RED_STDERR]:-}❌  $fatal_error_message${COLOUR[OFF_STDERR]:-}"
+        echo -e "Suggest: ${COLOUR[VIVID_BLUE_STDOUT]:-}$CMD_AS_DISPLAY --help${COLOUR[OFF_STDOUT]:-}"
+    } >&2
+    exit 1
+}
+
+function do_exitWithHelp()
 {
     local fatal_error_message="${1:-}"
 
     [[ -n "$fatal_error_message" ]] && echo -e "${COLOUR[VIVID_RED_STDERR]:-}❌  $fatal_error_message${COLOUR[OFF_STDERR]:-}" >&2
-
-    COLOUR[VIVID_BLUE_HELP]="${COLOUR[VIVID_BLUE_STDOUT]:-}"
-    COLOUR[VIVID_RED_HELP]="${COLOUR[VIVID_RED_STDOUT]:-}"
-    COLOUR[OFF_HELP]="${COLOUR[OFF_STDOUT]:-}"
-
     {
+
+        colours_setUsed 'STDOUT'
+
         [[ -n "${APP_DESCRIPTION:-}" ]] && echo -e "$CMD_AS_DISPLAY: $APP_DESCRIPTION"
         if declare -F app_help >/dev/null 2>&1 ; then
             app_help
             echo ""
             echo -e "Additional functions:"
         else
-            echo -e "Usage: ${COLOUR[VIVID_BLUE_STDOUT]:-}$CMD_AS_DISPLAY <parameters>${COLOUR[OFF_STDOUT]:-}"
+            echo -e "Usage: ${COLOUR[VIVID_BLUE_USED]:-}$CMD_AS_DISPLAY <parameters>${COLOUR[OFF_USED]:-}"
             echo -e ""
             echo -e "Parameters:"
         fi
         echo -e "     --help     : Give this help message"
-        [[ -n "${APP_VERSION:-}" ]] && echo -e "     --version  : Give version : ${COLOUR[VIVID_BLUE_STDOUT]:-}$APP_VERSION${COLOUR[OFF_STDOUT]:-}"
+        [[ -n "${APP_VERSION:-}" ]] && echo -e "     --version  : Give version : ${COLOUR[VIVID_BLUE_USED]:-}$APP_VERSION${COLOUR[OFF_USED]:-}"
 
         _installDirShow="$(displayPath "$INSTALL_DIR")"
-        [[ ":$PATH:" == *":${INSTALL_DIR}:"* ]] || _installDirShow+="${COLOUR[VIVID_RED_STDOUT]:-}  ⚠️  This should be added to \$PATH${COLOUR[OFF_STDOUT]:-}"
-        if [[ "$INSTALLATION_NOTE" == "OK" ]] ; then
-            echo -e "     --uninstall: Uninstalls from $_installDirShow"
-        else
-            echo -e "     --install  : Installs to $_installDirShow "
+        [[ ":$PATH:" == *":${INSTALL_DIR}:"* ]] || _installDirShow+="${COLOUR[VIVID_RED_USED]:-}  ⚠️  This should be added to \$PATH${COLOUR[OFF_USED]:-}"
+        if [[ "${APP_SELF_INSTALL:-}" == "yes" ]] ; then
+            if [[ "$INSTALLATION_NOTE" == "OK" ]] ; then
+                echo -e "     --uninstall: Uninstalls from $_installDirShow"
+            else
+                echo -e "     --install  : Installs to $_installDirShow "
+            fi
         fi
         echo -e "     --colours=no|yes|auto  (Default 'auto')"
         echo ""
@@ -163,10 +171,14 @@ function app_load_param_validate_from_list()
 {
     local name="$1"
     local value="$2"
-    local validation="$3"
-    echo "Validating value for $name=$value against expected values: $validation"
 
-    [[ ",$validation," != *",${value},"* ]] || do_exit_with_help "Invalid value for $name=$value Expected one of [$validation]"
+    shift 2 || true
+    #echo "ℹ️  Validating value for $name$value against expected values: [$(asCsvList "$@")]"
+
+    for arg in "$@"; do
+        [[ "$arg" == "$value" ]] && return 0
+    done
+    do_errorExitWithSuggestion "Invalid value for ${name}<value>.  <value> is ${value@Q}, but expected one of [$(asCsvList "$@")]"
 }
 
 function load_params()
@@ -178,13 +190,9 @@ function load_params()
     }
     function _app_get_param_option_name_value()
     {
-        declare -F app_load_param_option_name_value  &> /dev/null  && app_load_param_option_name_value "$1" "$2"
+        declare -F app_load_param_option_name_value  &> /dev/null  && app_load_param_option_name_value "$1" "${2:-}"
     }
 
-    function _app_get_param_option_name_only()
-    {
-        declare -F app_load_param_option_name_only  &> /dev/null  && app_load_param_option_name_only "$1"
-    }
 
     function _app_get_param_direct_value()
     {
@@ -194,6 +202,7 @@ function load_params()
 
     _app_get_param_defaults
     local giveCodeHelp='no'
+    local giveStdHelp='no'
     for arg in "$@"; do
         [[ -z "$arg" ]] && continue
         if [[ "$am_processing_options" == "yes" ]] && [[ "$arg" == '-'* ]]; then
@@ -203,30 +212,34 @@ function load_params()
                 echo "$(basename "$THIS_EXE") $APP_VERSION"
                 exit 0
             elif [[ "$arg" == '--help' ]]; then
-                do_exit_with_help
+                giveStdHelp='yes'
             elif [[ "$arg" == '--colours=no' ]] || [[ "$arg" == '--colours=yes' ]] || [[ "$arg" == '--colours=auto' ]]; then
                 colours_load "${arg#--colours=}"
-            elif [[ "$INSTALLATION_NOTE" == "OK" ]] && [[ "$arg" == "--uninstall" ]] ; then
-                (sleep 0.5 && rm -f "$THIS_EXE" && echo "Removed $THIS_EXE") &
-                exit 0
-            elif [[ "$INSTALLATION_NOTE" != "OK" ]] && [[ "$arg" == "--install" ]] ; then
-                do_install_directly
-                exit 0
+            elif [[ "${APP_SELF_INSTALL:-}" == "yes" ]] ; then
+                if [[ "$INSTALLATION_NOTE" == "OK" ]] && [[ "$arg" == "--uninstall" ]] ; then
+                    (sleep 0.5 && rm -f "$THIS_EXE" && echo "Removed $THIS_EXE") &
+                    exit 0
+                elif [[ "$arg" == "--install" ]] ; then # Always permit 'install' even if the app is already installed, as this can be used to fix an installation that is not working (e.g. due to missing copy or missing PATH entry)
+                    do_install_directly
+                    exit 0
+                fi
             elif [[ "$arg" == '--code-help' ]] ; then
                 giveCodeHelp='yes'
             elif [[ "$arg" == "--"*"="* ]] ; then
-                _app_get_param_option_name_value "${arg%%=*}" "${arg#*=}" || do_exit_with_help "Unknown named option ${arg@Q}"
+                _app_get_param_option_name_value "${arg%%=*}=" "${arg#*=}" || do_errorExitWithSuggestion "Unknown named option ${arg@Q}"
             else
-                _app_get_param_option_name_only  "${arg}"                 || do_exit_with_help "Unknown option ${arg@Q}"
+                _app_get_param_option_name_value "${arg}"                  || do_errorExitWithSuggestion "Unknown option ${arg@Q}"
             fi
         else
-            _app_get_param_direct_value "$arg" || do_exit_with_help "Unknown direct argument ${arg@Q}"
+            _app_get_param_direct_value "$arg" || do_errorExitWithSuggestion "Unknown direct argument ${arg@Q}"
         fi
     done
 
     if [[ "$giveCodeHelp" == "yes" ]]; then
         giveCodeHelp
         exit 0
+    elif [[ "$giveStdHelp" == "yes" ]]; then
+        do_exitWithHelp ""
     fi
 
 }
@@ -289,7 +302,6 @@ function giveWarning()
 }
 
 
-THIS_EXE="$(readlink -m "${BASH_SOURCE[0]}")"
 EXE_NAME="$(basename "$THIS_EXE")"
 app_name_and_version="${EXE_NAME}"
 [[ "${APP_VERSION:-}" ]] && app_name_and_version+=" ($APP_VERSION)"
@@ -306,8 +318,7 @@ else
 fi
 
 
-
-function coloursShow()
+function colours_show()
 {
 
         echo '╭─────────────────────────────────────────────────────────'
